+++ design-patterns--facade-patterns/README.md
@@ -1,17 +1,1 @@
 # Design Patterns: Facade Pattern
-
-## Problem Description
-
-We want to generate a `CHANGELOG` for release.
-
-### MVP
-
-Command line application that takes arguments: `public repository` and `previous version`
-
-App grabs data from GitHub and 
-
- takes in a public repository,
-a previously tagged release,
-
-
-Generate a CHANGELOG consisting of commit messages
