+++ design-patterns--facade-patterns/changelog/b_function_encapsulation.py
